✨ feat(planning): show level/stage badge on agenda cards, widen pinch capture
HITL feedback on the semantic zoom slice:

- Agenda is the full-detail level, so three-row cards now carry a badge
  with the match's level, stage and stage item next to the start time.
- Pinch/ctrl+wheel handling moves from the grid's scroll container to a
  wrapper around the whole planning content, so a pinch that starts next
  to the grid zooms the schedule instead of the browser page. Safari's
  proprietary gesture events are cancelled there too.
